Version 1.2
Finished the Split Feature. There are still some problems with the
presentation on IE9 but it displays correctly on Firefox and Chrome.

Commented out the footer as it needs to be upgraded to a "sticky footer"
which will cling to the bottom of the browser window.

Modified the rules page.